refactor: Split problem.py into problem/ subpackage (problem + constraint)
